W3D2/Q2/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Pillow==11.2.1
requests==2.32.4
google-generativeai==0.7.2

